Update TestGeneric1.java
Before going to the collection, First we look at Generics in java.
A generic  class is a class that is type-safe. This means a generic class can act upon any data type. Similarly, a generic interface is also  type-safe and hence it can use any data type. Generic classes and interfaces are also called 'parameterized types' because they use a parameter that determines which data type they should work upon.